+++ S2ZonaConfigurator/Enums/AppMode.cs
@@ -1,0 +1,7 @@
+﻿namespace S2ZonaConfigurator.Enums;
+
+public enum AppMode
+{
+    Main,
+    PakModsDiff
+}