+++ S2ZonaConfigurator/Interfaces/IAppService.cs
@@ -1,0 +1,5 @@
+﻿namespace S2ZonaConfigurator.Interfaces;
+public interface IAppService
+{
+    Task RunAsync();
+}