+++ S2ZonaConfigurator/Models/AppConfig.cs
@@ -1,4 +1,6 @@
-﻿namespace S2ZonaConfigurator.Models;
+﻿using S2ZonaConfigurator.Enums;
+
+namespace S2ZonaConfigurator.Models;
 
 public class AppConfig
 {
@@ -14,6 +16,7 @@ public class PathsConfig
     public string VanillaDirectory { get; set; } = "vanilla";
     public string ModifiedDirectory { get; set; } = "modified";
     public string ModsDirectory { get; set; } = "mods";
+    public string PakModsDirectory { get; set; } = "pak_mods";
     public string PaksPath { get; set; } = "Stalker2\\Content\\Paks";
     public string OutputPakName { get; set; } = "ZonaBundle.pak";
 }
@@ -26,6 +29,7 @@ public class GameConfig
 
 public class OptionsConfig
 {
+    public AppMode AppMode { get; set; } = AppMode.Main;
     public bool OutputChangelogFile { get; set; } = false;
     public bool CleanWorkDirectory { get; set; } = true;
     public bool GenerateDiffReport { get; set; } = false;